gcp/functions/daily-report/requirements.txt
google-cloud-bigquery>=3.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
